Fixed errors, Renamed most funcs from new to create
- Rename new to create:
But not those in class Element, only SourceParser
It might be a good idea to have separate names for those
creation functions in Element and SourceParser.
Now they're different, but also the ones in Element should
have some other name than newElement().
- Also fixed some easy/wrong error reports.
rae_render is now down to 49 errors from 60.